move runtime state map into util file
