+++ src/vs/workbench/contrib/positronConsole/common/sessionDisplayUtils.ts
@@ -9,8 +9,9 @@ import { Codicon } from '../../../../base/common/codicons.js';
 import { ThemeIcon } from '../../../../base/common/themables.js';
 import { IModelService } from '../../../../editor/common/services/model.js';
 import { asCssVariable } from '../../../../platform/theme/common/colorUtils.js';
-import { LanguageRuntimeSessionMode } from '../../../services/languageRuntime/common/languageRuntimeService.js';
+import { LanguageRuntimeSessionMode, RuntimeState } from '../../../services/languageRuntime/common/languageRuntimeService.js';
 import { IRuntimeSessionDisplayInfo } from '../../../services/runtimeSession/common/runtimeSessionService.js';
+import { RuntimeStatus } from '../browser/components/runtimeStatus.js';
 import { POSITRON_QUARTO_ICON } from '../../../common/theme.js';
 import { isQuartoDocument } from '../../positronQuarto/common/positronQuartoConfig.js';
 
@@ -95,3 +96,20 @@ export function getSessionIconStyle(info: SessionIconInfo, modelService: IModelS
 	}
 	return undefined;
 }
+
+/**
+ * Maps a RuntimeState to a RuntimeStatus for display purposes.
+ */
+export const runtimeStateToRuntimeStatus: Record<RuntimeState, RuntimeStatus> = {
+	[RuntimeState.Uninitialized]: RuntimeStatus.Disconnected,
+	[RuntimeState.Initializing]: RuntimeStatus.Active,
+	[RuntimeState.Starting]: RuntimeStatus.Active,
+	[RuntimeState.Restarting]: RuntimeStatus.Active,
+	[RuntimeState.Ready]: RuntimeStatus.Idle,
+	[RuntimeState.Idle]: RuntimeStatus.Idle,
+	[RuntimeState.Busy]: RuntimeStatus.Active,
+	[RuntimeState.Interrupting]: RuntimeStatus.Active,
+	[RuntimeState.Exiting]: RuntimeStatus.Active,
+	[RuntimeState.Exited]: RuntimeStatus.Disconnected,
+	[RuntimeState.Offline]: RuntimeStatus.Disconnected,
+};
